+++ packages/actions/templates/index.ts
@@ -1,4 +1,5 @@
 import { AppBuild } from "./app-build/config";
 import { DeployFirebase } from "./deploy-firebase/config";
+import { PRCreateContent } from "./pr-create-content/config";
 
-export const ACTION_TEMPLATES = { AppBuild, DeployFirebase };
+export const ACTION_TEMPLATES = { AppBuild, DeployFirebase, PRCreateContent };
